# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Booking Thin Client API",
        "description": "REST API тонкого клиента для взаимодействия с внешним билетным шлюзом. Все данные проксируются от поставщика. База данных не используется.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Локальная разработка"
        },
        {
            "url": "https://yourdomain.com/api/v1",
            "description": "Продакшн"
        }
    ],
    "paths": {
        "/api/v1/events/{eventId}/book": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Забронировать места на событие",
                "description": "Бронирует указанные места для события. Требуется имя покупателя.",
                "operationId": "7c15716c74af91ef78e16047ce700cd7",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "ID события",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для бронирования",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Места успешно забронированы",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные данные (например, недоступные места)"
                    },
                    "404": {
                        "description": "Событие не найдено"
                    },
                    "502": {
                        "description": "Ошибка шлюза"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookingRequest": {
                "required": [
                    "seat_ids",
                    "customer_name"
                ],
                "properties": {
                    "seat_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 501
                        }
                    },
                    "customer_name": {
                        "type": "string",
                        "example": "Иван Петров"
                    }
                },
                "type": "object"
            },
            "BookingResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "booking_id": {
                        "type": "string",
                        "example": "BOOK-12345"
                    },
                    "message": {
                        "type": "string",
                        "example": "Места успешно забронированы"
                    }
                },
                "type": "object"
            }
        }
    },
...
}